implemented dropout layer support for reducing overfitting.
Now either a 'd' or 'D' can be specified in the model definition at
'layer_order' so that either nn.Dropout or spatial nn.Dropout2D layers
can be added for potentially improved network regularization. In
addition, the dropout probability can be specified by a new
'dropout_prob' parameter which can take a single float, a list or tuple
of floats to allow to specify the dropout probability for each
single convolution in the double convolution step.